[refactor] refactor code to find nextLadderNumber
