+++ src/test/java/ladder/LineTest.java
@@ -1,7 +1,9 @@
 package ladder;
 
+import ladder.domain.LadderNumber;
 import ladder.domain.Line;
 import ladder.domain.Point;
+import org.assertj.core.api.AssertionsForClassTypes;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -51,7 +53,6 @@ public class LineTest {
         );
     }
 
-
     @Test
     @DisplayName("램던 라인 생성후 가로줄 겹치느지 여부 테스트")
     void create_random() {
@@ -71,4 +72,43 @@ public class LineTest {
         }
         return false;
     }
+
+
+    @Test
+    @DisplayName("next시 왼쪽에 연결된 다음 사다리 있을시 이동 되는지 테스트")
+    void next_to_left() {
+        int number = 2;
+        LadderNumber currentLadderNumber = new LadderNumber(number);
+        Line line = Line.of(Arrays.asList(new Point(false), new Point(false), new Point(true)));
+        LadderNumber next = line.nextLadderNumber(currentLadderNumber);
+        LadderNumber leftLadderNumber = new LadderNumber(number -1);
+
+        AssertionsForClassTypes.assertThat(next).isEqualTo(leftLadderNumber);
+    }
+
+    @Test
+    @DisplayName("next시 오른쪽에 연결된 다음 사다리 있을시 이동 되는지 테스트")
+    void next_to_right() {
+        int number = 1;
+        LadderNumber currentLadderNumber = new LadderNumber(number);
+        Line line = Line.of(Arrays.asList(new Point(false), new Point(false), new Point(true)));
+        LadderNumber next = line.nextLadderNumber(currentLadderNumber);
+        LadderNumber rightLadderNumber = new LadderNumber(number + 1);
+
+        AssertionsForClassTypes.assertThat(next).isEqualTo(rightLadderNumber);
+    }
+
+
+    @Test
+    @DisplayName("next시 연결된 라인이 없을시 그대로 인지 테스트")
+    void next_same() {
+        int number = 2;
+        LadderNumber currentLadderNumber = new LadderNumber(number);
+        Line line = Line.of(Arrays.asList(new Point(false), new Point(false), new Point(false)));
+        LadderNumber next = line.nextLadderNumber(currentLadderNumber);
+        LadderNumber leftLadderNumber = new LadderNumber(number);
+
+        AssertionsForClassTypes.assertThat(next).isEqualTo(leftLadderNumber);
+    }
+
 }
